rename hash_table <- hashtable
