+++ hash_table/test_hash_table.py
@@ -1,6 +1,6 @@
 import unittest
 
-from hashtable import HashTable
+from hash_table import HashTable
 
 
 class TestHashTable(unittest.TestCase):
