+++ hash_table/test_hash_table_no_collisions.py
@@ -6,7 +6,7 @@ Does not collide with DJB2 or FNV-1-64. But could collide with other hashes.
 """
 
 import unittest
-from hashtable import HashTable
+from hash_table import HashTable
 
 
 class TestHashTable(unittest.TestCase):
